% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_games.R
\name{baseball_games}
\alias{baseball_games}
\title{Retrieve the available baseball games for a given league, team and season during a specific time interval.}
\usage{
baseball_games(league = "mlb", team_id = "mlb-bos",
  interval_type = "regularseason", season_id = "mlb-2016",
  status = "ended", verbose = TRUE)
}
\arguments{
\item{league}{character. The baseball league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. MLB is default.}

\item{team_id}{character.  The team id, can be in the form of the slug "mlb-bos".  Default is the Boston Red Sox, mlb-bos.}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is mlb-2016.}

\item{status}{character.  That status of the game.  Default is ended.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the baseball games for the specified league.
}
\description{
A function to retrieve all of the available baseball games for a specified team.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- baseball_games(league="mlb", team_id="mlb-bos")
}
}

