% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_players.R
\name{baseball_players}
\alias{baseball_players}
\title{Retrieve players for a given league, team and season during a specific time interval.}
\usage{
baseball_players(league = "mlb", team_id = "mlb-bos",
  interval_type = "regularseason", season_id = "mlb-2016", verbose = TRUE)
}
\arguments{
\item{league}{character. The hockey league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. MLB the default.}

\item{team_id}{character.  The team id, can be in the form of the slug "mlb-bos".  Default is the Boston Red Sox, mlb-bos.}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is mlb-2016.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the baseball players for the specified league and team.
}
\description{
A function to retrieve all of the players in baseball, for a given team during a current season and interval type.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- baseball_players(league="mlb", team_id="mlb-bos")
}
}

