% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseball_teams.R
\name{baseball_teams}
\alias{baseball_teams}
\title{Retrieve the available baseball teams in a given league}
\usage{
baseball_teams(league = "mlb", verbose = TRUE)
}
\arguments{
\item{league}{character. The baseball league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. MLB is default.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the baseball teams for the specified league.
}
\description{
A function to retrieve all of the available baseball teams for a specified league.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- baseball_teams(league="mlb")
}
}

