% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_games.R
\name{basketball_games}
\alias{basketball_games}
\title{Retrieve the available basketball games for a given league, team and season during a specific time interval.}
\usage{
basketball_games(league = "nba", team_id = "nba-bos",
  interval_type = "regularseason", season_id = "nba-2015-2016",
  status = "ended", verbose = TRUE)
}
\arguments{
\item{league}{character. The basketball league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. NBA is default.}

\item{team_id}{character.  The team id, can be in the form of the slug "nba-bos".  Default is the Boston Red Sox, nba-bos.}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is nba-2015-2016.}

\item{status}{character.  That status of the game.  Default is ended.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the basketball games for the specified league.
}
\description{
A function to retrieve all of the available basketball games for a specified team.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- basketball_games(league="nba", team_id="nba-bos")
}
}

