% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_injuries.R
\name{basketball_injuries}
\alias{basketball_injuries}
\title{Retrieve the available basketball injuries for a given league, team and season during a specific time interval.}
\usage{
basketball_injuries(league = "nba", team_id = "nba-bos",
  interval_type = "regularseason", season_id = "nba-2015-2016",
  verbose = TRUE)
}
\arguments{
\item{league}{character. The basketball league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. NBA is default.}

\item{team_id}{character.  The team id, can be in the form of the slug "nba-bos".  Default is the Boston Celtics, nba-bos.}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is nba-2015-2016.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the basketball injuries for the specified team.
}
\description{
A function to retrieve all of the available basketball injuries for a specified team.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- basketball_injuries(league="nba", team_id="nba-bos")
}
}

