% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_players.R
\name{basketball_players}
\alias{basketball_players}
\title{Retrieve players for a given league, team and season during a specific time interval.}
\usage{
basketball_players(league = "nba", team_id = "nba-bos",
  interval_type = "regularseason", season_id = "nba-2015-2016",
  verbose = TRUE)
}
\arguments{
\item{league}{character. The hockey league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. NBA the default.}

\item{team_id}{character.  The team id, can be in the form of the slug "nba-bos".  Default is the Boston Red Sox, nba-bos.}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is nba-2015-2016.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the baseball players for the specified league and team.
}
\description{
A function to retrieve all of the players in baseball, for a given team during a current season and interval type.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- basketball_players(league="nba", team_id="nba-bos")
}
}

