% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball_teams.R
\name{basketball_teams}
\alias{basketball_teams}
\title{Retrieve the available basketball teams in a given league}
\usage{
basketball_teams(league = "nba", verbose = TRUE)
}
\arguments{
\item{league}{character. The basketball league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. NBA is default.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the basketball teams for the specified league.
}
\description{
A function to retrieve all of the available basketball teams for a specified league.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- basketball_teams(league="nba")
}
}

