% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sideload_conferences.R
\name{clean_sideload_conferences}
\alias{clean_sideload_conferences}
\title{Clean up the sideloaded conference data for easy merging with other sources of data.}
\usage{
clean_sideload_conferences(x)
}
\value{
a dataframe with the modified data cleaned up for merging
}
\description{
Clean the conferences sideloaded data for merging and reporting purposes
}
\examples{
\dontrun{
set_token("insert-your-token-here")
teams <- ss_get_result(sport = "hockey", league = "nhl", ep = "teams")
confs <- parse_stattle(teams, "conferences")
confs <- clean_sideload_conferences(confs)
}
}

