% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sideload_divisions.R
\name{clean_sideload_divisions}
\alias{clean_sideload_divisions}
\title{Clean up the sideloaded division data for easy merging with other sources of data.}
\usage{
clean_sideload_divisions(x, prefix = "division", drop_conf = FALSE)
}
\value{
a dataframe with the modified data cleaned up for merging
}
\description{
Clean the divisions sideloaded data for merging and reporting purposes
}
\examples{
\dontrun{
set_token("insert-your-token-here")
teams <- ss_get_result(sport = "hockey", league = "nhl", ep = "teams")
divs <- parse_stattle(teams, "divisions")
divs <- clean_sideload_divisions(divs)
}
}

