% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sideload_games.R
\name{clean_sideload_games}
\alias{clean_sideload_games}
\title{Clean up the sideloaded game data for easy merging with other sources of data.}
\usage{
clean_sideload_games(games, prefix = "game")
}
\value{
a dataframe with the modified data cleaned up for merging
}
\description{
Clean the game sideloaded data for merging and reporting purposes
}

