% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sideload_leagues.R
\name{clean_sideload_leagues}
\alias{clean_sideload_leagues}
\title{Clean up the sideloaded league data for easy merging with other sources of data.}
\usage{
clean_sideload_leagues(lgs)
}
\value{
a dataframe with the modified data cleaned up for merging
}
\description{
Clean the league sideloaded data for merging and reporting purposes
}
\examples{
\dontrun{
set_token("insert-your-token-here")
teams <- ss_get_result(sport = "hockey", league = "nhl", ep = "teams")
lgs <- parse_stattle(teams, "leagues")
teams <- clean_sideload_leagues(lgs)
}
}

