% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sideload_players.R
\name{clean_sideload_players}
\alias{clean_sideload_players}
\title{Clean up the sideloaded player data for easy merging with other sources of data.}
\usage{
clean_sideload_players(x, prefix = "player")
}
\value{
a dataframe with the modified data cleaned up for merging
}
\description{
Clean the teams sideloaded data for merging and reporting purposes
}
\examples{
\dontrun{
set_token("insert-your-token-here")
result <- ss_get_result(sport = "hockey", league = "nhl", ep = "injuries")
players <- parse_stattle(result, "players")
players <- clean_sideload_players(players)
}
}

