% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sideload_positions.R
\name{clean_sideload_positions}
\alias{clean_sideload_positions}
\title{Clean up the sideloaded position data for easy merging with other sources of data.}
\usage{
clean_sideload_positions(x, prefix = "position")
}
\value{
a dataframe with the modified data cleaned up for merging
}
\description{
Clean the position data sideloaded data for merging and reporting purposes
}

