% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sideload_teams.R
\name{clean_sideload_teams}
\alias{clean_sideload_teams}
\title{Clean up the sideloaded team data for easy merging with other sources of data.}
\usage{
clean_sideload_teams(x, prefix = "team")
}
\value{
a dataframe with the modified data cleaned up for merging
}
\description{
Clean the teams sideloaded data for merging and reporting purposes
}
\examples{
\dontrun{
set_token("insert-your-token-here")
result <- ss_get_result(sport = "hockey", league = "nhl", ep = "teams")
teams <- parse_stattle(result, "teams")
teams <- clean_sideload_teams(teams)
}
}

