% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sideload_venues.R
\name{clean_sideload_venues}
\alias{clean_sideload_venues}
\title{Clean up the sideloaded venue data for easy merging with other sources of data.}
\usage{
clean_sideload_venues(venues, prefix = "game_venue")
}
\value{
a dataframe with the modified data cleaned up for merging
}
\description{
Clean the venue sideloaded data for merging and reporting purposes
}

