% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_games.R
\name{football_games}
\alias{football_games}
\title{Retrieve the available football games for a given league, team and season during a specific time interval.}
\usage{
football_games(league = "nfl", team_id = "nfl-ne",
  interval_type = "regularseason", season_id = "nfl-2015",
  status = "ended", verbose = TRUE)
}
\arguments{
\item{league}{character. The football league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. nfl is default.}

\item{team_id}{character.  The team id, can be in the form of the slug "nfl-ne".  Default is the Boston Red Sox, nfl-ne.}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is nfl-2015.}

\item{status}{character.  That status of the game.  Default is ended.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the football games for the specified league.
}
\description{
A function to retrieve all of the available football games for a specified team.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- football_games(league="nfl", team_id="nfl-ne")
}
}

