% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_injuries.R
\name{football_injuries}
\alias{football_injuries}
\title{Retrieve the available football injuries for a given league, team and season during a specific time interval.}
\usage{
football_injuries(league = "nfl", team_id = "nfl-ne",
  interval_type = "regularseason", season_id = "nfl-2016", verbose = TRUE)
}
\arguments{
\item{league}{character. The football league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. nfl is default.}

\item{team_id}{character.  The team id, can be in the form of the slug "nfl-ne".  Default is the New England Patriots, nfl-ne.}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is nfl-2015.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the football injuries for the specified team.
}
\description{
A function to retrieve all of the available football injuries for a specified team.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- football_injuries(league="nfl", team_id="nfl-ne")
}
}

