% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_players.R
\name{football_players}
\alias{football_players}
\title{Retrieve players for a given league, team and season during a specific time interval.}
\usage{
football_players(league = "nfl", team_id = "nfl-ne",
  interval_type = "regularseason", season_id = "nfl-2015", verbose = TRUE)
}
\arguments{
\item{league}{character. The hockey league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. nfl the default.}

\item{team_id}{character.  The team id, can be in the form of the slug "nfl-ne".  Default is the Boston Patriots, nfl-ne.}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is nfl-2015.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the football players for the specified league and team.
}
\description{
A function to retrieve all of the players in football, for a given team during a current season and interval type.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- football_players(league="nfl", team_id="nfl-ne")
}
}

