% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football_teams.R
\name{football_teams}
\alias{football_teams}
\title{Retrieve the available football teams in a given league}
\usage{
football_teams(league = "nfl", verbose = TRUE)
}
\arguments{
\item{league}{character. The football league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. nfl is default.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the football teams for the specified league.
}
\description{
A function to retrieve all of the available football teams for a specified league.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- football_teams(league="nfl")
}
}

