% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_injuries.R
\name{hockey_injuries}
\alias{hockey_injuries}
\title{Retrieve hockey injuries for a league}
\usage{
hockey_injuries(league = "nhl", team_id = "", verbose = TRUE)
}
\arguments{
\item{league}{character. The hockey league to retrieve.  Currently only the NHL is supported, and is the default.}

\item{team_id}{character.  Optional. The team id, can be in the form of the slug "nhl-bos".  Default is all teams (empty character).}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the hockey injuries for the specified league, and optionally, for a given team.
}
\description{
A function to retrieve all of the injuries in hockey, for a given team.
}
\examples{
\dontrun{
setToken("insert-your-token-here")
results_bos <- hockey_injuries(league="nhl", team_id="nhl-bos") ## bruins injuries
results_all <- hockey_injuries(league="nhl") ## all injuries
}
}

