% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_teams.R
\name{hockey_teams}
\alias{hockey_teams}
\title{Retrieve the available hockey teams}
\usage{
hockey_teams(league = "nhl", verbose = TRUE)
}
\arguments{
\item{league}{character. The hockey league to retrieve.  Currently only the NHL is supported, and is the default.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the hockey teams for the specified league.
}
\description{
A function to retrieve all of the available hockey teams for a specified league.
}
\examples{
\dontrun{
setToken("insert-your-token-here")
results <- hockey_teams(league="nhl")
}
}

