% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_stattle.R
\name{parse_stattle}
\alias{parse_stattle}
\title{Utility function to quickly rbind all dataframes of an entry within the list returned by ss_get_result}
\usage{
parse_stattle(stattle_list, entry)
}
\arguments{
\item{stattle_list}{list A list of data, usually the object that is returned from ?ss_get_result}

\item{entry}{character. The entry from the API request which houses dataframes of data}
}
\value{
dataframe.
}
\description{
A helper to quickly bind dataframes
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- ss_get_result(sport="hockey", 
                         league="nhl",
                         ep = "feats",
                         query = list(game_id="nhl-2015-2016-nyr-bos-2015-11-27-1800"),
                         version = 1,
                         walk = FALSE,
                         page = NA,
                         verbose = TRUE)
results_df <- parse_stattle(results, "feats")
}
}

