% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_token.R
\name{set_token}
\alias{set_token}
\title{Set the Stattleship API token}
\usage{
set_token(token)
}
\arguments{
\item{token}{character. A valid Stattleship API token}
}
\description{
A simple function to set the Stattleship API token in an R environment variable
}
\examples{
\dontrun{
set_token("insert-your-token-here")}
}

