% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setToken.R
\name{setToken}
\alias{setToken}
\title{Set the Stattleship API token}
\usage{
setToken(token)
}
\arguments{
\item{token}{character. A valid Stattleship API token}
}
\description{
A simple function to set the Stattleship API token in an R environment variable
}
\examples{
\dontrun{
setToken("insert-your-token-here")}
}

