% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_data.R
\name{show_datasets}
\alias{show_datasets}
\title{Utility function to list the side-loaded data that accompanies a successful call with ss_get_result}
\usage{
show_datasets(stattle_list)
}
\arguments{
\item{stattle_list}{list A list of data the results from a valid call using ss_get_result}
}
\value{
vector  The elements that can be called during parse_stattle.
}
\description{
A helper to list the dataframes available to you as a result of a call with ss_gest_result
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- ss_get_result(sport="hockey", 
                         league="nhl",
                         ep = "feats",
                         query = list(game_id="nhl-2015-2016-nyr-bos-2015-11-27-1800"),
                         version = 1,
                         walk = FALSE,
                         page = NA,
                         verbose = TRUE)
results_df <- parse_stattle(results, "feats")
show_datasets(results_df)

}
}

