% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get_injuries.R, R/ss_get_teams.R
\name{ss_get_teams}
\alias{ss_get_teams}
\title{Retrieve the available teams}
\usage{
ss_get_teams(sport = "hockey", league = "nhl", verbose = TRUE)

ss_get_teams(sport = "hockey", league = "nhl", verbose = TRUE)
}
\arguments{
\item{sport}{character. The sport to retrieve.  Currently baseball, basketball, hockey, and football are supported.  hockey is default.}

\item{league}{character. The league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. NHL is default.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}

\item{sport}{character. The sport to retrieve.  Currently baseball, basketball, hockey, and football are supported.  hockey is default.}

\item{league}{character. The league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. NHL is default.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the teams for a given league and sport.

a dataframe of the teams for a given league and sport.
}
\description{
A function to retrieve all of the teams for a specified league and sport

A function to retrieve all of the teams for a specified league and sport
}
\examples{
\dontrun{
set_token("insert-your-token-here")
results <- ss_get_teams(sport="hockey", league="nhl")
}
\dontrun{
set_token("insert-your-token-here")
results <- ss_get_teams(sport="hockey", league="nhl")
}
}

