% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_keep_cols.R
\name{ss_keep_cols}
\alias{ss_keep_cols}
\title{Helper to only keep specified columns}
\usage{
ss_keep_cols(dat, cols = c())
}
\arguments{
\item{dat}{data.frame The original data frame}

\item{cols}{vector  A character vector of columns to keep.}
}
\value{
a dataframe with only the valid specified columns retained.
}
\description{
A helper function that takes a data frame and returns the same dataframe, but 
just with the specified columns.  This is helpfult to keep just the columns of interest 
and remove those applicable to other sports.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
teams = hockey_teams()
mycols = c("team_id", "team_hashtag", "team_name")
teams2 = ss_keep_cols(teams, mycols)
}
}

