% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_player_logs.R
\name{ss_player_logs}
\alias{ss_player_logs}
\title{Helper to get player's game logs for a specified team in a sport and league}
\usage{
ss_player_logs(sport = "hockey", league = "nhl", team_id = "nhl-bos",
  interval_type = "regularseason", season_id = "nhl-2015-2016",
  status = "ended", game_id = "", player_id = "", verbose = TRUE,
  walk = TRUE)
}
\arguments{
\item{sport}{character. The sport of the team}

\item{league}{character. The hockey league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. 
NHL is default.}

\item{team_id}{character.  The team id, can be in the form of the slug "nhl-bos".
Default is all teams (empty character).}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is nhl-2015-2016.}

\item{status}{character.  The status of the game.  Ended is default.}

\item{game_id}{character.  A specific game of interest.}

\item{player_id}{character.  A specific player of interest.}

\item{since}{character. A relative time range. e.g. 3 days ago.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}

\item{walk}{logical  Whether or not you want to get additional pages.  Default is TRUE.}
}
\value{
a dataframe of the player game logs
}
\description{
A function to interface and make it easier to get player game logs for a team in a sport/league and 
optionally, a game id.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
bos_logs <- ss_player_logs(league="nhl", team_id="nhl-bos") 
}
}

