% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_team_logs.R
\name{ss_team_logs}
\alias{ss_team_logs}
\title{Helper to get team game logs for a specified team in a sport and league}
\usage{
ss_team_logs(sport = "hockey", league = "nhl", team_id = "nhl-bos",
  interval_type = "regularseason", season_id = "nhl-2015-2016",
  status = "ended", game_id = "", verbose = TRUE, walk = TRUE)
}
\arguments{
\item{sport}{character. The sport of the team}

\item{league}{character. The hockey league to retrieve.  Currently MLB, NBA, NHL, and MLB are supported. 
NHL is default.}

\item{team_id}{character.  Optional. The team id, can be in the form of the slug "nhl-bos".
Default is all teams (empty character).}

\item{interval_type}{character.  The season interval.  Default is regularseason.}

\item{season_id}{character.  The season.  Default is nhl-2015-2016.}

\item{status}{character.  The status of the game.  Ended is default.}

\item{game_id}{character.  Optional. A specific game of interest.}

\item{since}{character. A relative time range. e.g. 3 days ago.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}

\item{walk}{logical.  TRUE will retrieve all of the pages. Default is TRUE.}
}
\value{
a dataframe of the teams game logs
}
\description{
A function to interface and make it easier to get team game logs for a team in a sport/league.
}
\examples{
\dontrun{
set_token("insert-your-token-here")
bos_logs <- ss_team_logs(league="nhl", team_id="nhl-bos") 
}
}

