% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattle.R
\name{stattle}
\alias{stattle}
\title{Interface with the Stattleship API}
\usage{
stattle(token, sport = "hockey", league = "nhl", ep = "stats",
  query = list(), version = 1, walk = F, page = NA, verbose = F)
}
\arguments{
\item{sport}{character. The sport, such as hockey, basketball, football}

\item{league}{character. NHL, NBA, NFL, etc.}

\item{ep}{character. The endpoint}

\item{query}{A list that defines the query parameters}

\item{version}{The API version. Current version is 1.}

\item{walk}{logical. if TRUE, walks through and returns all results if there is more than one page of results}

\item{page}{numeric. The page number to request}

\item{verbose}{logical. For debugging, returns response and parsed response.}
}
\description{
A simple, generic function to query data from the Stattleship API
}
\examples{
\dontrun{
setToken("insert-your-token-here")
results <- stattle(sport="hockey", 
                   league="nhl",
                   ep = "stats",
                   query = list()
                   version = 1,
                   walk = TRUE,
                   page = NA,
                   verbose = FALSE)
}
}

