% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey_games.R
\name{hockey_games}
\alias{hockey_games}
\title{Retrieve the available hockey games for a given team}
\usage{
hockey_games(league = "nhl", team_id = "nhl-bos", verbose = TRUE)
}
\arguments{
\item{league}{character. The hockey league to retrieve.  Currently only the NHL is supported, and is the default.}

\item{team_id}{character.  The team id, can be in the form of the slug "nhl-bos".  Default is the Boston Bruins, nhl-bos.}

\item{verbose}{logical.  TRUE will print messages to the console.  Default is TRUE.}
}
\value{
a dataframe of the hockey teams for the specified league.
}
\description{
A function to retrieve all of the available hockey games for a specified team.
}
\examples{
\dontrun{
setToken("insert-your-token-here")
results <- hockey_games(league="nhl", team_id="nhl-bos")
}
}

