% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGetResult.R
\name{ssGetResult}
\alias{ssGetResult}
\title{Interface with the Stattleship API}
\usage{
ssGetResult(token, sport = "hockey", league = "nhl", ep = "teams",
  query = list(), version = 1, walk = FALSE, page = NA,
  verbose = TRUE)
}
\arguments{
\item{sport}{character. The sport, such as hockey, basketball, football. Default is hockey.}

\item{league}{character. NHL, NBA, NFL, etc. Default is nhl.}

\item{ep}{character. The endpoint.  Default is teams.}

\item{query}{list. A list that defines the query parameters. Default is empty list.}

\item{version}{numeric. The API version. Current version is 1 and is the default value.}

\item{walk}{logical. If TRUE, walks through and returns all results if there is more than one page of results.  Default is FALSE.}

\item{page}{numeric. The page number to request.  Default is NA.}

\item{verbose}{logical. For debugging, prints status messages to the console, which can be helpful for walking through results. Default is TRUE.}
}
\value{
a list of lists.  If `walk=FALSE`, it will be a list of length 1.  If `walk=TRUE`, a list of lists may be returned depending on how many pages are returned.
}
\description{
A simple, generic function to query data from the Stattleship API
}
\examples{
\dontrun{
setToken("insert-your-token-here")
results <- ssGetResult(sport="hockey", 
                       league="nhl",
                       ep = "teams",
                       query = list()
                       version = 1,
                       walk = FALSE,
                       page = NA,
                       verbose = TRUE)
}
}

